requirements.txt
streamlit==1.38.0
google-generativeai==0.8.2
requests==2.32.3
